+++ test/dummy/app/models/user.rb
@@ -1,7 +1,6 @@
 class User < ActiveRecord::Base
+  include ActiveModel::ForbiddenAttributesProtection
   include Authorizable::User  
-
-  attr_accessible :first_name, :last_name, :password, :password_confirmation, :email
 
   def access_roles
     [].tap do |allowed|